adds further spill planning functionality to modal
