+++ src/app/modules/core/services/nwisservices.service.ts
@@ -3,7 +3,8 @@ import { HttpClient, HttpHeaderResponse } from '@angular/common/http';
 import { BehaviorSubject, Observable, of, Subject } from 'rxjs';
 import {catchError, map} from 'rxjs/operators';
 //import { gages } from '../models/gages';
-import { MapService } from '../services/map.service'
+import { MapService } from '../services/map.service';
+import * as L from 'leaflet';
 
 @Injectable()
 export class NWISService {
@@ -93,5 +94,52 @@ export class NWISService {
         return of(false);
       }}
       ))
+  }
+
+  public check4gages(features) {
+    const layerGroup = new L.FeatureGroup([]);
+    const reportlayerGroup = new L.FeatureGroup([]);
+    let gagesArray = [];
+    features.forEach(i => {
+
+      if (i.geometry.type === 'Point') {
+        var siteID = i.properties.identifier.replace("USGS-", "");
+		// MarkerMaker Icon
+		var blackTriangle = L.divIcon({className: 'wmm-triangle wmm-black wmm-size-15'});
+        layerGroup.addLayer(L.marker([i.geometry.coordinates[1], i.geometry.coordinates[0]], { 
+			icon: blackTriangle
+		}).bindPopup('<h3>Streamgages</h3><br /><b>Station ID: </b>' + siteID + '<br /><b>Station Name: </b>' + i.properties.name + '<br /><b>Station Latitude: </b>' + i.geometry.coordinates[1] + '<br /><b>Station Longitude: </b>' + i.geometry.coordinates[0] + '<br /><b>NWIS page: </b><a href="' + i.properties.uri + '" target="_blank">link</a><br /><b>StreamStats Gage page: </b><a href="https://streamstatsags.cr.usgs.gov/gagepages/html/' + siteID + '.htm" target="_blank">link</a>'));
+        reportlayerGroup.addLayer(L.marker([i.geometry.coordinates[1], i.geometry.coordinates[0]], { 
+			icon: blackTriangle
+		}));
+        gagesArray.push(i);
+      } 
+    })
+
+    //check if there is a gage data;
+    if (gagesArray.length > 0) {
+      for (var i = 0; i < gagesArray.length; i++) {
+        features.forEach(o => {
+          if (o.geometry.type !== "Point") {
+            if (gagesArray[i].properties.comid == String(o.properties.nhdplus_comid)) {
+              gagesArray[i].properties["drainagearea"] = o.properties.DrainageArea / 2.59; //in sqmiles
+            } else { }
+          }
+        })
+      };
+      //create service
+      //add gage
+      this.gagesArray.next(gagesArray);
+      this.MapService.gageDischargeSearch.next(true);
+      this.getMostRecentFlow(gagesArray);
+    } else {
+      this.MapService.gageDischargeSearch.next(true);
+    };
+
+    // because it is async it takes time to process function above, once we have it done - we get the bounds
+    // Potential to improve
+    setTimeout(() => {
+      this.MapService.setBounds(layerGroup.getBounds());
+    });
   }
 }
